# annotrieve-api-specs.yaml
openapi: 3.1.0
info:
  version: "1.0.0"
  title: "Annotrieve API"
  description: "API for retrieving and managing eukaryotic genome annotations"
  contact:
    email: "emilio.righi@crg.eu"
servers:
  - url: https://genome.crg.es/annotrieve/api/v0
    description: Production server
...
paths:
  /annotations:
    get:
      tags:
        - "annotations"
      operationId: "getAnnotations"
      summary: "Get annotations"
      description: "Returns annotation metadata filtered, sorted, and paginated via query parameters."
      parameters:
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/taxids"
        - $ref: "#/components/parameters/assembly_accessions"
        - $ref: "#/components/parameters/bioproject_accessions"
        - $ref: "#/components/parameters/db_sources"
        - $ref: "#/components/parameters/feature_sources"
        - $ref: "#/components/parameters/biotypes"
        - $ref: "#/components/parameters/feature_types"
        - $ref: "#/components/parameters/pipelines"
        - $ref: "#/components/parameters/providers"
        - $ref: "#/components/parameters/md5_checksums"
        - $ref: "#/components/parameters/has_stats"
        - $ref: "#/components/parameters/refseq_categories"
        - $ref: "#/components/parameters/assembly_levels"
        - $ref: "#/components/parameters/assembly_statuses"
        - $ref: "#/components/parameters/assembly_types"
        - $ref: "#/components/parameters/release_date_from"
        - $ref: "#/components/parameters/release_date_to"
        - $ref: "#/components/parameters/sort_by"
        - $ref: "#/components/parameters/sort_order"
        - $ref: "#/components/parameters/fields"
      responses:
        "200":
          description: "Annotations page"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAnnotationsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - "annotations"
      operationId: "postAnnotations"
      summary: "Get annotations via POST"
      description: "Same as GET /annotations, but accepts filters in the request body."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationQueryParams"
      responses:
        "200":
          description: "Annotations page"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAnnotationsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /annotations/report:
    get:
      tags:
        - "annotations"
      operationId: "getAnnotationsReport"
      summary: "Download annotation metadata as TSV"
      description: "Streams a TSV file with annotation metadata columns (annotation_id, assembly_accession, assembly_name, organism_name, taxid, database, provider, source_url, bgzip_path, csi_path)."
      parameters:
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/taxids"
        - $ref: "#/components/parameters/assembly_accessions"
        - $ref: "#/components/parameters/bioproject_accessions"
        - $ref: "#/components/parameters/db_sources"
        - $ref: "#/components/parameters/feature_sources"
        - $ref: "#/components/parameters/biotypes"
        - $ref: "#/components/parameters/feature_types"
        - $ref: "#/components/parameters/pipelines"
        - $ref: "#/components/parameters/providers"
        - $ref: "#/components/parameters/md5_checksums"
        - $ref: "#/components/parameters/has_stats"
        - $ref: "#/components/parameters/refseq_categories"
        - $ref: "#/components/parameters/assembly_levels"
        - $ref: "#/components/parameters/assembly_statuses"
        - $ref: "#/components/parameters/assembly_types"
        - $ref: "#/components/parameters/release_date_from"
        - $ref: "#/components/parameters/release_date_to"
        - $ref: "#/components/parameters/sort_by"
        - $ref: "#/components/parameters/sort_order"
      responses:
        "200":
          description: "TSV stream"
          content:
            text/tab-separated-values:
              schema:
                type: string
                description: "TSV payload streamed line by line"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - "annotations"
      operationId: "postAnnotationsReport"
      summary: "Download annotation metadata via POST"
      description: "Same as GET /annotations/report, but accepts filters in the request body."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationQueryParams"
      responses:
        "200":
          description: "TSV stream"
          content:
            text/tab-separated-values:
              schema:
                type: string
                description: "TSV payload streamed line by line"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /annotations/stats/summary:
    get:
      tags:
        - "annotations"
      operationId: "getAnnotationsSummaryStats"
      summary: "Get annotation summary statistics"
      description: "Aggregates gene, transcript, and feature statistics for the filtered annotation set."
      parameters:
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/taxids"
        - $ref: "#/components/parameters/assembly_accessions"
        - $ref: "#/components/parameters/bioproject_accessions"
        - $ref: "#/components/parameters/db_sources"
        - $ref: "#/components/parameters/feature_sources"
        - $ref: "#/components/parameters/biotypes"
        - $ref: "#/components/parameters/feature_types"
        - $ref: "#/components/parameters/pipelines"
        - $ref: "#/components/parameters/providers"
        - $ref: "#/components/parameters/md5_checksums"
        - $ref: "#/components/parameters/has_stats"
        - $ref: "#/components/parameters/refseq_categories"
        - $ref: "#/components/parameters/assembly_levels"
        - $ref: "#/components/parameters/assembly_statuses"
        - $ref: "#/components/parameters/assembly_types"
        - $ref: "#/components/parameters/release_date_from"
        - $ref: "#/components/parameters/release_date_to"
        - $ref: "#/components/parameters/sort_by"
        - $ref: "#/components/parameters/sort_order"
      responses:
        "200":
          description: "Summary statistics"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnotationsSummaryStats"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - "annotations"
      operationId: "postAnnotationsSummaryStats"
      summary: "Get annotation summary statistics via POST"
      description: "Same as GET /annotations/stats/summary, but accepts filters in the request body."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationQueryParams"
      responses:
        "200":
          description: "Summary statistics"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnotationsSummaryStats"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /annotations/frequencies:
    get:
      tags:
        - "annotations"
      operationId: "getAnnotationFrequencyFields"
      summary: "List allowed annotation frequency fields"
      responses:
        "200":
          description: "Frequency-capable fields"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrequencyFieldList"
        "500":
          $ref: "#/components/responses/InternalError"
  /annotations/frequencies/{field}:
    get:
      tags:
        - "annotations"
      operationId: "getAnnotationFrequencies"
      summary: "Get annotation field frequencies"
      description: "Returns frequency counts for the selected field across filtered annotations."
      parameters:
        - name: field
          in: path
          required: true
          description: "Field to aggregate. Allowed values: taxid, organism_name, database, release_date, assembly_accession, assembly_name, feature_type, feature_source, biotype, pipeline, provider."
          schema:
            type: string
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/taxids"
        - $ref: "#/components/parameters/assembly_accessions"
        - $ref: "#/components/parameters/bioproject_accessions"
        - $ref: "#/components/parameters/db_sources"
        - $ref: "#/components/parameters/feature_sources"
        - $ref: "#/components/parameters/biotypes"
        - $ref: "#/components/parameters/feature_types"
        - $ref: "#/components/parameters/pipelines"
        - $ref: "#/components/parameters/providers"
        - $ref: "#/components/parameters/md5_checksums"
        - $ref: "#/components/parameters/has_stats"
        - $ref: "#/components/parameters/refseq_categories"
        - $ref: "#/components/parameters/assembly_levels"
        - $ref: "#/components/parameters/assembly_statuses"
        - $ref: "#/components/parameters/assembly_types"
        - $ref: "#/components/parameters/release_date_from"
        - $ref: "#/components/parameters/release_date_to"
      responses:
        "200":
          description: "Frequency counts"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrequencyCounts"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - "annotations"
      operationId: "postAnnotationFrequencies"
      summary: "Get annotation field frequencies via POST"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationQueryParams"
      responses:
        "200":
          description: "Frequency counts"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrequencyCounts"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /annotations/errors:
    get:
      tags:
        - "annotations"
      operationId: "getAnnotationErrors"
      summary: "List annotation import errors"
      parameters:
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: "Annotation errors"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAnnotationErrorsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /annotations/{md5_checksum}:
    get:
      tags:
        - "annotations"
      operationId: "getAnnotationById"
      summary: "Get annotation by checksum"
      parameters:
        - name: md5_checksum
          in: path
          required: true
          description: "MD5 checksum of the annotation (32 lowercase hex characters)."
          schema:
            type: string
            pattern: '^[a-f0-9]{32}$'
      responses:
        "200":
          description: "Annotation details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Annotation"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /annotations/{md5_checksum}/gff:
    get:
      tags:
        - "annotations"
      operationId: "streamAnnotationGff"
      summary: "Stream filtered GFF3 for an annotation"
      description: "Streams GFF3 lines for the given annotation. At least one filter (region, feature_type, feature_source, or biotype) must be provided."
      parameters:
        - name: md5_checksum
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-f0-9]{32}$'
        - name: region
          in: query
          description: "Sequence/contig identifier or alias."
          schema:
            type: string
        - name: start
          in: query
          description: "Start position (0-based)."
          schema:
            type: integer
            minimum: 0
        - name: end
          in: query
          description: "End position (inclusive)."
          schema:
            type: integer
            minimum: 0
        - name: feature_type
          in: query
          description: "Filter by GFF feature type (column 3)."
          schema:
            type: string
        - name: feature_source
          in: query
          description: "Filter by GFF feature source (column 2)."
          schema:
            type: string
        - name: biotype
          in: query
          description: "Filter by feature biotype (attribute)."
          schema:
            type: string
      responses:
        "200":
          description: "GFF3 stream"
          content:
            text/plain:
              schema:
                type: string
                description: "GFF3 lines streamed as plain text"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /annotations/{md5_checksum}/contigs:
    get:
      tags:
        - "annotations"
      operationId: "getAnnotationContigs"
      summary: "List contigs available for an annotation"
      description: "Streams newline-separated contig identifiers from the annotation's Tabix index."
      parameters:
        - name: md5_checksum
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-f0-9]{32}$'
      responses:
        "200":
          description: "Contig IDs"
          content:
            text/plain:
              schema:
                type: string
                description: "Newline-separated contig identifiers"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /annotations/{md5_checksum}/contigs/aliases:
    get:
      tags:
        - "annotations"
      operationId: "getAnnotationContigAliases"
      summary: "Get contig alias mappings"
      parameters:
        - name: md5_checksum
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-f0-9]{32}$'
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: "Sequence alias mappings"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSequenceAliasesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    filter:
      name: "filter"
      in: "query"
      description: "Text search filter across multiple fields"
      schema:
        type: "string"
    limit:
      name: "limit"
      in: "query"
      description: "Maximum number of results to return"
      schema:
        type: "integer"
        minimum: 1
        maximum: 1000
        default: 20
    offset:
      name: "offset"
      in: "query"
      description: "Number of results to skip"
      schema:
        type: "integer"
        minimum: 0
        default: 0
    taxids:
      name: "taxids"
      in: "query"
      description: "Comma-separated list of taxonomy IDs"
      schema:
        type: "string"
      example: "9606,10090,7227"
    bioproject_accessions:
      name: "bioproject_accessions"
      in: "query"
      description: "Comma-separated list of BioProject accessions"
      schema:
        type: "string"
      example: "PRJNA123456,PRJNA123457"
    assembly_accessions:
      name: "assembly_accessions"
      in: "query"
      description: "Comma-separated list of assembly accessions"
      schema:
        type: "string"
      example: "GCA_000001405.29,GCF_000001405.40"
    db_sources:
      name: "db_sources"
      in: "query"
      description: "Comma-separated list of source databases (RefSeq, Ensembl, GenBank)"
      schema:
        type: "string"
    feature_sources:
      name: "feature_sources"
      in: "query"
      description: "Comma-separated list of GFF feature sources (column 2)"
      schema:
        type: "string"
    biotypes:
      name: "biotypes"
      in: "query"
      description: "Comma-separated list of feature biotypes"
      schema:
        type: "string"
    feature_types:
      name: "feature_types"
      in: "query"
      description: "Comma-separated list of GFF feature types (column 3)"
      schema:
        type: "string"
    pipelines:
      name: "pipelines"
      in: "query"
      description: "Comma-separated list of annotation pipeline names"
      schema:
        type: "string"
    providers:
      name: "providers"
      in: "query"
      description: "Comma-separated list of annotation providers"
      schema:
        type: "string"
    md5_checksums:
      name: "md5_checksums"
      in: "query"
      description: "Comma-separated list of annotation MD5 checksums"
      schema:
        type: "string"
    has_stats:
      name: "has_stats"
      in: "query"
      description: "Filter annotations that have computed statistics"
      schema:
        type: "boolean"
...
    refseq_categories:
      name: "refseq_categories"
      in: "query"
      description: "Comma-separated list of RefSeq categories"
      schema:
        type: "string"
    assembly_levels:
      name: "assembly_levels"
      in: "query"
      description: "Comma-separated list of assembly levels"
      schema:
        type: "string"
    assembly_statuses:
      name: "assembly_statuses"
      in: "query"
      description: "Comma-separated list of assembly statuses"
      schema:
        type: "string"
    assembly_types:
      name: "assembly_types"
      in: "query"
      description: "Comma-separated list of assembly types"
      schema:
        type: "string"
...
    release_date_from:
      name: "release_date_from"
      in: "query"
      description: "Filter annotations released on or after this ISO date"
      schema:
        type: "string"
        format: "date"
    release_date_to:
      name: "release_date_to"
      in: "query"
      description: "Filter annotations released on or before this ISO date"
      schema:
        type: "string"
        format: "date"
    sort_by:
      name: "sort_by"
      in: "query"
      description: "Field to sort results by"
      schema:
        type: "string"
    sort_order:
      name: "sort_order"
      in: "query"
      description: "Sort order"
      schema:
        type: "string"
        enum: ["asc", "desc"]
    fields:
      name: "fields"
      in: "query"
      description: "Comma-separated list of dot-notation fields to include in the response"
      schema:
        type: "string"
...
  responses:
    BadRequest:
      description: "Bad request"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: "Resource not found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: "Internal server error"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
      required:
        - detail
    AnnotationQueryParams:
      type: object
      properties:
        filter:
          type: string
        taxids:
          type: string
        bioproject_accessions:
          type: string
        assembly_accessions:
          type: string
        db_sources:
          type: string
        feature_sources:
          type: string
        biotypes:
          type: string
        feature_types:
          type: string
        pipelines:
          type: string
        providers:
          type: string
        md5_checksums:
          type: string
        has_stats:
          type: boolean
        latest_release_by:
          type: string
          enum: ["organism", "assembly"]
        refseq_categories:
          type: string
        assembly_levels:
          type: string
        assembly_statuses:
          type: string
        assembly_types:
          type: string
        release_date_from:
          type: string
          format: date
        release_date_to:
          type: string
          format: date
        sort_by:
          type: string
        sort_order:
          type: string
          enum: ["asc", "desc"]
        offset:
          type: integer
          default: 0
        limit:
          type: integer
          default: 20
        fields:
          type: string
...
    Annotation:
      type: object
      properties:
        annotation_id:
          type: string
          description: "MD5 checksum of the annotation"
        assembly_accession:
          type: string
        assembly_name:
          type: string
        organism_name:
          type: string
        taxid:
          type: string
        taxon_lineage:
          type: array
          items:
            type: string
        mapped_regions:
          type: array
          items:
            type: string
        source_file_info:
          $ref: "#/components/schemas/SourceFileInfo"
        indexed_file_info:
          $ref: "#/components/schemas/IndexedFileInfo"
        features_summary:
          $ref: "#/components/schemas/FeatureSummary"
        features_statistics:
          $ref: "#/components/schemas/GFFStats"
      required:
        - annotation_id
        - assembly_accession
        - assembly_name
        - organism_name
        - taxid
...
    AnnotationsSummaryStats:
      type: object
      properties:
        annotations:
          type: object
          properties:
            total_count:
              type: integer
            related_organisms_count:
              type: integer
            related_assemblies_count:
              type: integer
        genes:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/GeneAggregateStats"
        transcripts:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/TranscriptAggregateStats"
        features:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/FeatureAggregateStats"
    GeneAggregateStats:
      type: object
      properties:
        total_count:
          type: integer
        mean_count:
          type: number
        median_count:
          type: number
        mean_length:
          type: number
        median_length:
          type: number
    TranscriptAggregateStats:
      type: object
      properties:
        total_count:
          type: integer
        mean_count:
          type: number
        median_count:
          type: number
        mean_length:
          type: number
        median_length:
          type: number
    FeatureAggregateStats:
      type: object
      properties:
        mean_length:
          type: number
        median_length:
          type: number
    AnnotationError:
      type: object
      properties:
        assembly_accession:
          type: string
        taxid:
          type: string
        organism_name:
          type: string
        error_message:
          type: string
        url_path:
          type: string
          format: uri
        source_md5:
          type: string
        source_database:
          type: string
        release_date:
          type: string
          format: date-time
        last_modified:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
      required:
        - assembly_accession
        - taxid
        - organism_name
        - error_message
        - url_path
        - source_md5
        - source_database
...
    FrequencyFieldList:
      type: object
      properties:
        fields:
          type: array
          items:
            type: string
    FrequencyCounts:
      type: object
      additionalProperties:
        type: integer
...
    PaginatedAnnotationsResponse:
      type: object
      properties:
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/Annotation"
...
    PaginatedAnnotationErrorsResponse:
      type: object
      properties:
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/AnnotationError"
    PaginatedSequenceAliasesResponse:
      type: object
      properties:
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/SequenceAlias"
...